+++ src/resource.ts
@@ -1,6 +1,6 @@
-import * as path from "path";
 import {
   Command,
+  FileDecoration,
   SourceControlResourceDecorations,
   SourceControlResourceState,
   ThemeColor,
@@ -10,39 +10,7 @@ import { PropStatus, Status } from "./common/types";
 import { memoize } from "./decorators";
 import { configuration } from "./helpers/configuration";
 
-// Path needs to be relative from out/
-const iconsRootPath = path.join(__dirname, "..", "icons");
-
-function getIconUri(iconName: string, theme: string): Uri {
-  return Uri.file(path.join(iconsRootPath, theme, `${iconName}.svg`));
-}
-
 export class Resource implements SourceControlResourceState {
-  private static icons: any = {
-    light: {
-      Added: getIconUri("status-added", "light"),
-      Conflicted: getIconUri("status-conflicted", "light"),
-      Deleted: getIconUri("status-deleted", "light"),
-      Ignored: getIconUri("status-ignored", "light"),
-      Missing: getIconUri("status-missing", "light"),
-      Modified: getIconUri("status-modified", "light"),
-      Renamed: getIconUri("status-renamed", "light"),
-      Replaced: getIconUri("status-replaced", "light"),
-      Unversioned: getIconUri("status-unversioned", "light")
-    },
-    dark: {
-      Added: getIconUri("status-added", "dark"),
-      Conflicted: getIconUri("status-conflicted", "dark"),
-      Deleted: getIconUri("status-deleted", "dark"),
-      Ignored: getIconUri("status-ignored", "dark"),
-      Missing: getIconUri("status-missing", "dark"),
-      Modified: getIconUri("status-modified", "dark"),
-      Renamed: getIconUri("status-renamed", "dark"),
-      Replaced: getIconUri("status-replaced", "dark"),
-      Unversioned: getIconUri("status-unversioned", "dark")
-    }
-  };
-
   constructor(
     private _resourceUri: Uri,
     private _type: string,
@@ -72,9 +40,6 @@ export class Resource implements SourceControlResourceState {
   }
 
   get decorations(): SourceControlResourceDecorations {
-    // TODO@joh, still requires restart/redraw in the SCM viewlet
-    const light = { iconPath: this.getIconPath("light") };
-    const dark = { iconPath: this.getIconPath("dark") };
     const tooltip = this.tooltip;
     const strikeThrough = this.strikeThrough;
     const faded = this.faded;
@@ -82,9 +47,7 @@ export class Resource implements SourceControlResourceState {
     return {
       strikeThrough,
       faded,
-      tooltip,
-      light,
-      dark
+      tooltip
     };
   }
 
@@ -117,20 +80,6 @@ export class Resource implements SourceControlResourceState {
       title: "Open Diff With Base",
       arguments: [this]
     };
-  }
-
-  private getIconPath(theme: string): Uri | undefined {
-    if (this.type === Status.ADDED && this.renameResourceUri) {
-      return Resource.icons[theme].Renamed;
-    }
-
-    const type = this.type.charAt(0).toUpperCase() + this.type.slice(1);
-
-    if (typeof Resource.icons[theme][type] !== "undefined") {
-      return Resource.icons[theme][type];
-    }
-
-    return void 0;
   }
 
   private get tooltip(): string {
@@ -226,5 +175,11 @@ export class Resource implements SourceControlResourceState {
       default:
         return 1;
     }
+  }
+
+  get resourceDecoration(): FileDecoration {
+    const res = new FileDecoration(this.letter, this.tooltip, this.color);
+    res.propagate = this.type !== Status.DELETED;
+    return res;
   }
 }